chore(utils): move utils/types.h to kernel/types.h
