+++ include/kernel/arch/i686/gdt.h
@@ -24,8 +24,9 @@
 #ifndef KERNEL_ARCH_I686_GDT_H
 #define KERNEL_ARCH_I686_GDT_H
 
+#include <kernel/types.h>
+
 #include <utils/compiler.h>
-#include <utils/types.h>
 
 // Hard coded constant linear base 32-bits address for the TSS
 #define GDT_TSS_BASE_ADDRESS 0x00000000UL
